requirements.txt
asgiref==3.11.0
dj-database-url==3.1.2
Django==6.0.1
djangorestframework==3.16.1
gunicorn==25.1.0
packaging==26.0
sqlparse==0.5.5
tzdata==2025.3
